Revert to check native tars to pinpoint incompatibilities
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@33594 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/test/test_tar.sh
@@ -1,12 +1,13 @@
 #! /bin/sh
 
-# Favor GNU tar if available, as vendor versions may have issues with
-# corner cases.
-if gtar --version >/dev/null 2>&1; then
-  tar=gtar
+# Locate native tar;  exit if none found
+string="`whereis tar 2>/dev/null`"
+if [ -z "$string" ]; then
+  tar="`which tar 2>/dev/null`"
 else
-  tar=tar
+  tar="`echo $string | cut -f2 -d' '`"
 fi
+test -x "$tar"  ||  exit 0
 
 mkdir /tmp/test_tar.$$.1  ||  exit 1
 trap 'rm -rf /tmp/test_tar.$$.* &' 0 1 2 15
